Split ext/OceananigansMakieExt.jl into 4 scripts
