+++ ext/OceananigansMakieExt/OceananigansMakieExt.jl
@@ -1,0 +1,11 @@
+module OceananigansMakieExt
+
+include("DimensionalityUtils.jl")
+include("MakieConversions.jl")
+include("PlotExtensions.jl")
+
+using .DimensionalityUtils
+using .MakieConversions
+using .PlotExtensions
+
+end # module OceananigansMakieExt
